pyproject.toml
[tool.poetry]
name = "split_text_bot"
version = "0.0.1"
description = ""
authors = ["Nikita Churikov <nikita@chur.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
aiogram = "^2.8"
loguru = "^0.5.0"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
bot = "split_text_bot.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
